update(web): update labels to compatible with new tag schema
